Move Repository tests to a separate file.
